'OpenIDConnectClient.php5' renamed to 'OpenIDConnectClient.php'
